tools/requirements_ml.txt
ultralytics
onnx
onnxruntime
openvino-dev
psutil
opencv-python
numpy
